+++ packages/itwin/grouping-mapping-widget/src/widget/components/MappingAction.tsx
@@ -62,9 +62,9 @@ const MappingAction = ({ iModelId, mapping, returnFn }: MappingActionProps) => {
         title={mapping ? "Modify Mapping" : "Add Mapping"}
         returnFn={returnFn}
       />
-      <div className='details-form-container'>
-        <Fieldset legend='Mapping Details' className='details-form'>
-          <Small className='field-legend'>
+      <div className='gmw-details-form-container'>
+        <Fieldset legend='Mapping Details' className='gmw-details-form'>
+          <Small className='gmw-field-legend'>
             Asterisk * indicates mandatory fields.
           </Small>
           <LabeledInput